requirements.txt
# requirements.txt
# Core utilities
docutils
markdown
# For RAG
beautifulsoup4
sentence-transformers
faiss-cpu
# For local LLM interaction
openai
python-dotenv
